Style: fix the code format
